+++ app/src/ui/mod.rs
@@ -31,6 +31,7 @@ pub mod templates;
 pub mod thread_list;
 pub mod thread_row;
 pub mod translation;
+pub mod unsubscribe;
 pub mod vacation;
 pub mod welcome;
 pub mod when;